Refactored service tree item class
